requirements.txt
gradio 
fpdf 
python-docx
huggingface-hub==0.24.6 